+++ src/components/Test/Test.scss
@@ -4,3 +4,14 @@
     justify-content: center;
     padding: 0 5%;
 }
+
+.icon {
+    width: 36px; /* Ширина иконки */
+    height: 36px; /* Высота иконки */
+    vertical-align: middle; /* Выравнивание иконки по вертикали */
+    margin-right: 5px; /* Отступ между иконкой и текстом */
+}
+
+.icon.active {
+    fill: #ccc; /* Цвет SVG-изображения в активном состоянии */
+}
